@@ -1,9 +1,9 @@
 module github.com/sei-protocol/sei-chain
 
-go 1.19
+go 1.21
 
 require (
-	github.com/BurntSushi/toml v1.1.0
+	github.com/BurntSushi/toml v1.3.2
 	github.com/CosmWasm/wasmd v0.27.0
 	github.com/CosmWasm/wasmvm v1.0.1
 	github.com/armon/go-metrics v0.4.1
