@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bytes"
 	"encoding/json"
 	"flag"
 	"fmt"
@@ -17,21 +18,23 @@ import (
 	"syscall"
 	"time"
 
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	tokenfactorytypes "github.com/sei-protocol/sei-chain/x/tokenfactory/types"
-
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-	"github.com/cosmos/cosmos-sdk/std"
-
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/types"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	"github.com/cosmos/cosmos-sdk/std"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/tx"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	distributiontypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	"github.com/ethereum/go-ethereum/common"
+	"golang.org/x/sync/semaphore"
+	"golang.org/x/time/rate"
+
 	"github.com/sei-protocol/sei-chain/app"
 	dextypes "github.com/sei-protocol/sei-chain/x/dex/types"
+	tokenfactorytypes "github.com/sei-protocol/sei-chain/x/tokenfactory/types"
 )
 
 var TestConfig EncodingConfig
@@ -40,7 +43,12 @@ const (
 	VortexData = "{\"position_effect\":\"Open\",\"leverage\":\"1\"}"
 )
 
-var FromMili = sdk.NewDec(1000000)
+var (
+	FromMili      = sdk.NewDec(1000000)
+	producedCount = atomic.Int64{}
+	sentCount     = atomic.Int64{}
+	prevSentCount = atomic.Int64{}
+)
 
 type BlockData struct {
 	Txs []string `json:"txs"`
@@ -65,14 +73,46 @@ func init() {
 	std.RegisterInterfaces(TestConfig.InterfaceRegistry)
 	app.ModuleBasics.RegisterLegacyAminoCodec(TestConfig.Amino)
 	app.ModuleBasics.RegisterInterfaces(TestConfig.InterfaceRegistry)
+	// Add this so that we don't end up getting disconnected for EVM client
+	http.DefaultTransport.(*http.Transport).MaxIdleConnsPerHost = 100
 }
 
-func run(config Config) {
+// deployEvmContract executes a bash script and returns its output as a string.
+//
+//nolint:gosec
+func deployEvmContract(scriptPath string, config *Config) (common.Address, error) {
+	cmd := exec.Command(scriptPath, config.EVMRpcEndpoint())
+	var out bytes.Buffer
+	cmd.Stdout = &out
+	err := cmd.Run()
+	if err != nil {
+		return common.Address{}, err
+	}
+	return common.HexToAddress(out.String()), nil
+}
+
+func deployEvmContracts(config *Config) {
+	config.EVMAddresses = &EVMAddresses{}
+	if config.ContainsAnyMessageTypes(ERC20) {
+		fmt.Println("Deploying ERC20 contract")
+		erc20, err := deployEvmContract("loadtest/contracts/deploy_erc20.sh", config)
+		if err != nil {
+			fmt.Println("error deploying, make sure 0xF87A299e6bC7bEba58dbBe5a5Aa21d49bCD16D52 is funded")
+			panic(err)
+		}
+		config.EVMAddresses = &EVMAddresses{
+			ERC20: erc20,
+		}
+	}
+}
+
+func run(config *Config) {
 	// Start metrics collector in another thread
 	metricsServer := MetricsServer{}
-	go metricsServer.StartMetricsClient(config)
+	go metricsServer.StartMetricsClient(*config)
 
-	startLoadtestWorkers(config)
+	deployEvmContracts(config)
+	startLoadtestWorkers(*config)
 }
 
 // starts loadtest workers. If config.Constant is true, then we don't gather loadtest results and let producer/consumer
@@ -81,48 +121,41 @@ func startLoadtestWorkers(config Config) {
 	fmt.Printf("Starting loadtest workers\n")
 	client := NewLoadTestClient(config)
 	client.SetValidators()
-
 	configString, _ := json.Marshal(config)
 	fmt.Printf("Running with \n %s \n", string(configString))
-
-	txQueue := make(chan []byte, 10000)
-	done := make(chan struct{})
-	numProducers := 1000
-	var wg sync.WaitGroup
 
 	// Catch OS signals for graceful shutdown
 	signals := make(chan os.Signal, 1)
 	signal.Notify(signals, syscall.SIGINT, syscall.SIGTERM)
 
-	ticker := time.NewTicker(1 * time.Second)
-	start := time.Now()
-	var producedCount int64 = 0
-	var sentCount int64 = 0
-	var prevSentCount int64 = 0
 	var blockHeights []int
 	var blockTimes []string
 	var startHeight = getLastHeight(config.BlockchainEndpoint)
-	fmt.Printf("Starting loadtest producers\n")
-	// preload all accounts
-	keys := client.SignerClient.GetTestAccountsKeys(int(config.TargetTps))
-	for i := 0; i < numProducers; i++ {
-		wg.Add(1)
-		go client.BuildTxs(txQueue, i, keys, &wg, done, &producedCount)
+	keys := client.AccountKeys
+
+	// Create producers and consumers
+	fmt.Printf("Starting loadtest producers and consumers\n")
+	txQueues := make([]chan SignedTx, len(keys))
+	for i := range txQueues {
+		txQueues[i] = make(chan SignedTx, 10)
 	}
-	// Give producers some time to populate queue
-	if config.TargetTps > 1000 {
-		time.Sleep(5 * time.Second)
+	done := make(chan struct{})
+	producerRateLimiter := rate.NewLimiter(rate.Limit(config.TargetTps), int(config.TargetTps))
+	consumerSemaphore := semaphore.NewWeighted(int64(config.TargetTps))
+	var wg sync.WaitGroup
+	for i := 0; i < len(keys); i++ {
+		go client.BuildTxs(txQueues[i], i, &wg, done, producerRateLimiter, &producedCount)
+		go client.SendTxs(txQueues[i], i, done, &sentCount, consumerSemaphore, &wg)
 	}
 
-	fmt.Printf("Starting loadtest consumers\n")
-	go client.SendTxs(txQueue, done, &sentCount, int(config.TargetTps), &wg)
 	// Statistics reporting goroutine
+	ticker := time.NewTicker(10 * time.Second)
 	go func() {
+		start := time.Now()
 		for {
 			select {
 			case <-ticker.C:
 				currHeight := getLastHeight(config.BlockchainEndpoint)
-
 				for i := startHeight; i <= currHeight; i++ {
 					_, blockTime, err := getTxBlockInfo(config.BlockchainEndpoint, strconv.Itoa(i))
 					if err != nil {
@@ -133,11 +166,14 @@ func startLoadtestWorkers(config Config) {
 					blockTimes = append(blockTimes, blockTime)
 				}
 
-				printStats(start, &producedCount, &sentCount, &prevSentCount, blockHeights, blockTimes)
+				totalProduced := producedCount.Load()
+				totalSent := sentCount.Load()
+				prevTotalSent := prevSentCount.Load()
+				printStats(start, totalProduced, totalSent, prevTotalSent, blockHeights, blockTimes)
 				startHeight = currHeight
-				start = time.Now()
 				blockHeights, blockTimes = nil, nil
-
+				start = time.Now()
+				prevSentCount.Store(totalSent)
 			case <-done:
 				ticker.Stop()
 				return
@@ -147,19 +183,22 @@ func startLoadtestWorkers(config Config) {
 
 	// Wait for a termination signal
 	<-signals
-	fmt.Println("SIGINT received, shutting down...")
+	fmt.Println("SIGINT received, shutting down producers and consumers...")
 	close(done)
 
+	fmt.Println("Waiting for wait groups...")
+
 	wg.Wait()
-	close(txQueue)
+	fmt.Println("Closing channels...")
+	for i := range txQueues {
+		close(txQueues[i])
+	}
 }
 
-func printStats(startTime time.Time, producedCount *int64, sentCount *int64, prevSentCount *int64, blockHeights []int, blockTimes []string) {
+func printStats(startTime time.Time, totalProduced int64, totalSent int64, prevTotalSent int64, blockHeights []int, blockTimes []string) {
 	elapsed := time.Since(startTime)
-	produced := atomic.LoadInt64(producedCount)
-	sent := atomic.LoadInt64(sentCount)
-	tps := float64(sent-*prevSentCount) / elapsed.Seconds()
-	*prevSentCount = sent
+	tps := float64(totalSent-prevTotalSent) / elapsed.Seconds()
+
 	var totalDuration time.Duration
 	var prevTime time.Time
 	for i, blockTimeStr := range blockTimes {
@@ -175,16 +214,16 @@ func printStats(startTime time.Time, producedCount *int64, sentCount *int64, pre
 		fmt.Printf("Unable to calculate stats, not enough data. Skipping...\n")
 	} else {
 		avgDuration := totalDuration.Milliseconds() / int64(len(blockTimes)-1)
-		fmt.Printf("High Level - Time Elapsed: %v, Produced: %d, Sent: %d, TPS: %f, Avg Block Time: %d ms\nBlock Heights %v\n\n", elapsed, produced, sent, tps, avgDuration, blockHeights)
+		fmt.Printf("High Level - Time Elapsed: %v, Produced: %d, Sent: %d, TPS: %f, Avg Block Time: %d ms\nBlock Heights %v\n\n", elapsed, totalProduced, totalSent, tps, avgDuration, blockHeights)
 	}
 }
 
 // Generate a random message, only generate one admin message per block to prevent acc seq errors
 func (c *LoadTestClient) getRandomMessageType(messageTypes []string) string {
-	rand.Seed(time.Now().UnixNano())
-	messageType := messageTypes[rand.Intn(len(messageTypes))]
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	messageType := messageTypes[r.Intn(len(messageTypes))]
 	for c.generatedAdminMessageForBlock && c.isAdminMessageMapping[messageType] {
-		messageType = messageTypes[rand.Intn(len(messageTypes))]
+		messageType = messageTypes[r.Intn(len(messageTypes))]
 	}
 
 	if c.isAdminMessageMapping[messageType] {
@@ -193,25 +232,26 @@ func (c *LoadTestClient) getRandomMessageType(messageTypes []string) string {
 	return messageType
 }
 
-func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey, msgPerTx uint64) ([]sdk.Msg, bool, cryptotypes.PrivKey, uint64, int64) {
+func (c *LoadTestClient) generateMessage(key cryptotypes.PrivKey, msgType string) ([]sdk.Msg, bool, cryptotypes.PrivKey, uint64, int64) {
 	var msgs []sdk.Msg
-	messageType := c.getRandomMessageType(config.MessageTypes)
-
-	defer IncrTxMessageType(messageType)
-
+	config := c.LoadTestConfig
+	msgPerTx := config.MsgsPerTx
 	signer := key
 
-	defaultMessageTypeConfig := c.LoadTestConfig.PerMessageConfigs["default"]
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	defer IncrTxMessageType(msgType)
+
+	defaultMessageTypeConfig := config.PerMessageConfigs["default"]
 	gas := defaultMessageTypeConfig.Gas
 	fee := defaultMessageTypeConfig.Fee
 
-	messageTypeConfig, ok := c.LoadTestConfig.PerMessageConfigs[messageType]
+	messageTypeConfig, ok := config.PerMessageConfigs[msgType]
 	if ok {
 		gas = messageTypeConfig.Gas
 		fee = messageTypeConfig.Fee
 	}
 
-	switch messageType {
+	switch msgType {
 	case Vortex:
 		price := config.PriceDistr.Sample()
 		quantity := config.QuantityDistr.Sample()
@@ -245,6 +285,7 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 	case Bank:
 		msgs = []sdk.Msg{}
 		for i := 0; i < int(msgPerTx); i++ {
+
 			msgs = append(msgs, &banktypes.MsgSend{
 				FromAddress: sdk.AccAddress(key.PubKey().Address()).String(),
 				ToAddress:   sdk.AccAddress(key.PubKey().Address()).String(),
@@ -271,7 +312,7 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 	case CollectRewards:
 		adminKey := c.SignerClient.GetAdminKey()
 		delegatorAddr := sdk.AccAddress(adminKey.PubKey().Address())
-		operatorAddress := c.Validators[rand.Intn(len(c.Validators))].OperatorAddress
+		operatorAddress := c.Validators[r.Intn(len(c.Validators))].OperatorAddress
 		randomValidatorAddr, err := sdk.ValAddressFromBech32(operatorAddress)
 		if err != nil {
 			panic(err.Error())
@@ -301,9 +342,10 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 		}}
 	case Staking:
 		delegatorAddr := sdk.AccAddress(key.PubKey().Address()).String()
-		chosenValidator := c.Validators[rand.Intn(len(c.Validators))].OperatorAddress
+		chosenValidator := c.Validators[r.Intn(len(c.Validators))].OperatorAddress
 		// Randomly pick someone to redelegate / unbond from
 		srcAddr := ""
+		c.mtx.RLock()
 		for k := range c.DelegationMap[delegatorAddr] {
 			if k == chosenValidator {
 				continue
@@ -311,11 +353,12 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 			srcAddr = k
 			break
 		}
+		c.mtx.RUnlock()
 		msgs = []sdk.Msg{c.generateStakingMsg(delegatorAddr, chosenValidator, srcAddr)}
 	case Tokenfactory:
 		denomCreatorAddr := sdk.AccAddress(key.PubKey().Address()).String()
 		// No denoms, let's mint
-		randNum := rand.Float64()
+		randNum := r.Float64()
 		denom, ok := c.TokenFactoryDenomOwner[denomCreatorAddr]
 		switch {
 		case !ok || randNum <= 0.33:
@@ -339,7 +382,7 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 		}
 	case FailureBankMalformed:
 		var denom string
-		if rand.Float64() < 0.5 {
+		if r.Float64() < 0.5 {
 			denom = "unknown"
 		} else {
 			denom = "other"
@@ -384,7 +427,7 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 		contract := config.ContractDistr.Sample()
 		orderPlacements := generateDexOrderPlacements(config, key, msgPerTx, price, quantity)
 		var amountUsei int64
-		if rand.Float64() < 0.5 {
+		if r.Float64() < 0.5 {
 			amountUsei = 10000 * price.Mul(quantity).Ceil().RoundInt64()
 		} else {
 			amountUsei = 0
@@ -399,73 +442,30 @@ func (c *LoadTestClient) generateMessage(config Config, key cryptotypes.PrivKey,
 			ContractAddr: contract,
 			Funds:        amount,
 		}}
-	case WasmOccIteratorWrite:
-		// generate some values for indices 1-100
-		indices := []int{}
-		for i := 0; i < 100; i++ {
-			indices = append(indices, i)
-		}
-		rand.Shuffle(100, func(i, j int) {
-			indices[i], indices[j] = indices[j], indices[i]
-		})
-		values := [][]uint64{}
-		num_values := rand.Int()%100 + 1
-		for x := 0; x < num_values; x++ {
-			values = append(values, []uint64{uint64(indices[x]), rand.Uint64() % 12345})
-		}
-		contract := config.SeiTesterAddress
-		msgData := WasmIteratorWriteMsg{
-			Values: values,
-		}
-		jsonData, err := json.Marshal(msgData)
-		if err != nil {
-			panic(err)
-		}
-		msgs = []sdk.Msg{&wasmtypes.MsgExecuteContract{
-			Sender:   sdk.AccAddress(key.PubKey().Address()).String(),
-			Contract: contract,
-			Msg:      wasmtypes.RawContractMessage([]byte(fmt.Sprintf("{\"test_occ_iterator_write\":%s}", jsonData))),
-		}}
-	case WasmOccIteratorRange:
-		contract := config.SeiTesterAddress
-		start := rand.Uint32() % 100
-		end := rand.Uint32() % 100
-		if start > end {
-			start, end = end, start
-		}
-		msgs = []sdk.Msg{&wasmtypes.MsgExecuteContract{
-			Sender:   sdk.AccAddress(key.PubKey().Address()).String(),
-			Contract: contract,
-			Msg:      wasmtypes.RawContractMessage([]byte(fmt.Sprintf("{\"test_occ_iterator_range\":{\"start\": %d, \"end\": %d}}", start, end))),
-		}}
-	case WasmOccParallelWrite:
-		contract := config.SeiTesterAddress
-		// generate random value
-		value := rand.Uint64()
-		msgs = []sdk.Msg{&wasmtypes.MsgExecuteContract{
-			Sender:   sdk.AccAddress(key.PubKey().Address()).String(),
-			Contract: contract,
-			Msg:      wasmtypes.RawContractMessage([]byte(fmt.Sprintf("{\"test_occ_parallelism\":{\"value\": %d}}", value))),
-		}}
+
 	default:
-		fmt.Printf("Unrecognized message type %s\n", messageType)
+		fmt.Printf("Unrecognized message type %s", config.MessageType)
 	}
 
-	if strings.Contains(messageType, "failure") {
+	if strings.Contains(config.MessageType, "failure") {
 		return msgs, true, signer, gas, int64(fee)
 	}
 	return msgs, false, signer, gas, int64(fee)
 }
 
 func sampleDexOrderType(config Config) (orderType dextypes.OrderType) {
-	msgType := config.MsgTypeDistr.SampleDexMsgs()
-	switch msgType {
-	case Limit:
-		orderType = dextypes.OrderType_LIMIT
-	case Market:
-		orderType = dextypes.OrderType_MARKET
-	default:
-		panic(fmt.Sprintf("Unknown message type %s\n", msgType))
+	if config.MessageType == "failure_bank_malformed" {
+		orderType = -1
+	} else {
+		msgType := config.MsgTypeDistr.SampleDexMsgs()
+		switch msgType {
+		case Limit:
+			orderType = dextypes.OrderType_LIMIT
+		case Market:
+			orderType = dextypes.OrderType_MARKET
+		default:
+			panic(fmt.Sprintf("Unknown message type %s\n", msgType))
+		}
 	}
 	return orderType
 }
@@ -473,8 +473,9 @@ func sampleDexOrderType(config Config) (orderType dextypes.OrderType) {
 func generateDexOrderPlacements(config Config, key cryptotypes.PrivKey, msgPerTx uint64, price sdk.Dec, quantity sdk.Dec) (orderPlacements []*dextypes.Order) {
 	orderType := sampleDexOrderType(config)
 
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
 	var direction dextypes.PositionDirection
-	if rand.Float64() < 0.5 {
+	if r.Float64() < 0.5 {
 		direction = dextypes.PositionDirection_LONG
 	} else {
 		direction = dextypes.PositionDirection_SHORT
@@ -498,6 +499,8 @@ func generateDexOrderPlacements(config Config, key cryptotypes.PrivKey, msgPerTx
 }
 
 func (c *LoadTestClient) generateStakingMsg(delegatorAddr string, chosenValidator string, srcAddr string) sdk.Msg {
+	c.mtx.Lock()
+	defer c.mtx.Unlock()
 	// Randomly unbond, redelegate or delegate
 	// However, if there are no delegations, do so first
 	var msg sdk.Msg
@@ -540,9 +543,10 @@ func (c *LoadTestClient) generateVortexOrder(config Config, key cryptotypes.Priv
 	var msgs []sdk.Msg
 	contract := config.WasmMsgTypes.Vortex.ContractAddr
 
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
 	// Randomly select Position Direction
 	var direction dextypes.PositionDirection
-	if rand.Float64() < 0.5 {
+	if r.Float64() < 0.5 {
 		direction = dextypes.PositionDirection_LONG
 	} else {
 		direction = dextypes.PositionDirection_SHORT
@@ -666,5 +670,5 @@ func main() {
 
 	config := ReadConfig(*configFilePath)
 	fmt.Printf("Using config file: %s\n", *configFilePath)
-	run(config)
+	run(&config)
 }
