@@ -2,33 +2,34 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
 	"fmt"
-	"math"
 	"math/rand"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-	"golang.org/x/sync/semaphore"
 	"golang.org/x/time/rate"
-	"google.golang.org/grpc/connectivity"
 
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	"github.com/cosmos/cosmos-sdk/types"
 	typestx "github.com/cosmos/cosmos-sdk/types/tx"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-
-	"crypto/tls"
-
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"golang.org/x/sync/semaphore"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/connectivity"
 	"google.golang.org/grpc/credentials"
 )
 
 type LoadTestClient struct {
 	LoadTestConfig     Config
 	TestConfig         EncodingConfig
+	AccountKeys        []cryptotypes.PrivKey
 	TxClients          []typestx.ServiceClient
+	EvmTxClients       []*EvmTxClient
 	SignerClient       *SignerClient
 	ChainID            string
 	GrpcConns          []*grpc.ClientConn
@@ -43,11 +44,54 @@ type LoadTestClient struct {
 	generatedAdminMessageForBlock bool
 	// Messages that has to be sent from the admin account
 	isAdminMessageMapping map[string]bool
+	mtx                   *sync.RWMutex
 }
 
 func NewLoadTestClient(config Config) *LoadTestClient {
-	var dialOptions []grpc.DialOption
+	signerClient := NewSignerClient(config.NodeURI)
+	keys := signerClient.GetTestAccountsKeys(int(config.MaxAccounts))
+	txClients, grpcConns := BuildGrpcClients(config)
+	var evmTxClients []*EvmTxClient
+	if config.EvmRpcEndpoints != "" {
+		if config.ContainsAnyMessageTypes(EVM, ERC20) {
+			evmTxClients = BuildEvmTxClients(config, keys)
+		}
+	}
 
+	return &LoadTestClient{
+		LoadTestConfig:                config,
+		AccountKeys:                   keys,
+		TestConfig:                    TestConfig,
+		TxClients:                     txClients,
+		EvmTxClients:                  evmTxClients,
+		SignerClient:                  signerClient,
+		ChainID:                       config.ChainID,
+		GrpcConns:                     grpcConns,
+		StakingQueryClient:            stakingtypes.NewQueryClient(grpcConns[0]),
+		DelegationMap:                 map[string]map[string]int{},
+		TokenFactoryDenomOwner:        map[string]string{},
+		generatedAdminMessageForBlock: false,
+		isAdminMessageMapping:         map[string]bool{CollectRewards: true, DistributeRewards: true},
+		mtx:                           &sync.RWMutex{},
+	}
+}
+
+func (c *LoadTestClient) SetValidators() {
+	if strings.Contains(c.LoadTestConfig.MessageType, "staking") {
+		resp, err := c.StakingQueryClient.Validators(context.Background(), &stakingtypes.QueryValidatorsRequest{})
+		if err != nil {
+			panic(err)
+		}
+		c.Validators = resp.Validators
+	}
+}
+
+// BuildGrpcClients build a list of grpc clients
+func BuildGrpcClients(config Config) ([]typestx.ServiceClient, []*grpc.ClientConn) {
+	grpcEndpoints := strings.Split(config.GrpcEndpoints, ",")
+	txClients := make([]typestx.ServiceClient, len(grpcEndpoints))
+	grpcConns := make([]*grpc.ClientConn, len(grpcEndpoints))
+	var dialOptions []grpc.DialOption
 	dialOptions = append(dialOptions, grpc.WithDefaultCallOptions(
 		grpc.MaxCallRecvMsgSize(20*1024*1024),
 		grpc.MaxCallSendMsgSize(20*1024*1024)),
@@ -58,15 +102,12 @@ func NewLoadTestClient(config Config) *LoadTestClient {
 	} else {
 		dialOptions = append(dialOptions, grpc.WithInsecure())
 	}
-	endpoints := strings.Split(config.GrpcEndpoints, ",")
-	TxClients := make([]typestx.ServiceClient, len(endpoints))
-	GrpcConns := make([]*grpc.ClientConn, len(endpoints))
-	for i, endpoint := range endpoints {
+	for i, endpoint := range grpcEndpoints {
 		grpcConn, _ := grpc.Dial(
 			endpoint,
 			dialOptions...)
-		TxClients[i] = typestx.NewServiceClient(grpcConn)
-		GrpcConns[i] = grpcConn
+		txClients[i] = typestx.NewServiceClient(grpcConn)
+		grpcConns[i] = grpcConn
 		// spin up goroutine for monitoring and reconnect purposes
 		go func() {
 			for {
@@ -84,32 +125,39 @@ func NewLoadTestClient(config Config) *LoadTestClient {
 			}
 		}()
 	}
-
-	return &LoadTestClient{
-		LoadTestConfig:                config,
-		TestConfig:                    TestConfig,
-		TxClients:                     TxClients,
-		SignerClient:                  NewSignerClient(config.NodeURI),
-		ChainID:                       config.ChainID,
-		GrpcConns:                     GrpcConns,
-		StakingQueryClient:            stakingtypes.NewQueryClient(GrpcConns[0]),
-		DelegationMap:                 map[string]map[string]int{},
-		TokenFactoryDenomOwner:        map[string]string{},
-		generatedAdminMessageForBlock: false,
-		isAdminMessageMapping:         map[string]bool{CollectRewards: true, DistributeRewards: true},
-	}
+	return txClients, grpcConns
 }
 
-func (c *LoadTestClient) SetValidators() {
-	for _, mType := range c.LoadTestConfig.MessageTypes {
-		if mType == Staking {
-			resp, err := c.StakingQueryClient.Validators(context.Background(), &stakingtypes.QueryValidatorsRequest{})
-			if err != nil {
-				panic(err)
-			}
-			c.Validators = resp.Validators
-		}
+// BuildEvmTxClients build a list of EvmTxClients with a list of go-ethereum client
+func BuildEvmTxClients(config Config, keys []cryptotypes.PrivKey) []*EvmTxClient {
+	clients := make([]*EvmTxClient, len(keys))
+	ethEndpoints := strings.Split(config.EvmRpcEndpoints, ",")
+	if len(ethEndpoints) == 0 {
+		return clients
 	}
+	ethClients := make([]*ethclient.Client, len(ethEndpoints))
+	for i, endpoint := range ethEndpoints {
+		client, err := ethclient.Dial(endpoint)
+		if err != nil {
+			fmt.Printf("Failed to connect to endpoint %s with error %s", endpoint, err.Error())
+		}
+		ethClients[i] = client
+	}
+	// Get chainId
+	chainID, err := ethClients[0].NetworkID(context.Background())
+	if err != nil {
+		panic(fmt.Sprintf("Failed to get chain ID: %v \n", err))
+	}
+	// Get gas price
+	gasPrice, err := ethClients[0].SuggestGasPrice(context.Background())
+	if err != nil {
+		panic(fmt.Sprintf("Failed to suggest gas price: %v\n", err))
+	}
+	// Build one client per key
+	for i, key := range keys {
+		clients[i] = NewEvmTxClient(key, chainID, gasPrice, ethClients, config.EVMAddresses)
+	}
+	return clients
 }
 
 func (c *LoadTestClient) Close() {
@@ -118,77 +166,115 @@ func (c *LoadTestClient) Close() {
 	}
 }
 
-func (c *LoadTestClient) BuildTxs(txQueue chan<- []byte, producerId int, keys []cryptotypes.PrivKey, wg *sync.WaitGroup, done <-chan struct{}, producedCount *int64) {
+func (c *LoadTestClient) BuildTxs(
+	txQueue chan SignedTx,
+	keyIndex int,
+	wg *sync.WaitGroup,
+	done <-chan struct{},
+	rateLimiter *rate.Limiter,
+	producedCount *atomic.Int64,
+) {
+	wg.Add(1)
 	defer wg.Done()
 	config := c.LoadTestConfig
-
 	for {
 		select {
 		case <-done:
-			fmt.Printf("Stopping producer %d\n", producerId)
 			return
 		default:
-			key := keys[atomic.LoadInt64(producedCount)%int64(len(keys))]
-			msgs, _, _, gas, fee := c.generateMessage(config, key, config.MsgsPerTx)
-			txBuilder := TestConfig.TxConfig.NewTxBuilder()
-			_ = txBuilder.SetMsgs(msgs...)
-			txBuilder.SetGasLimit(gas)
-			txBuilder.SetFeeAmount([]types.Coin{
-				types.NewCoin("usei", types.NewInt(fee)),
-			})
-			// Use random seqno to get around txs that might already be seen in mempool
-			c.SignerClient.SignTx(c.ChainID, &txBuilder, key, uint64(rand.Intn(math.MaxInt)))
-			txBytes, _ := TestConfig.TxConfig.TxEncoder()(txBuilder.GetTx())
+			if !rateLimiter.Allow() {
+				continue
+			}
+			// Generate a message type first
+			messageTypes := strings.Split(config.MessageType, ",")
+			messageType := c.getRandomMessageType(messageTypes)
+			var signedTx SignedTx
+			// Sign EVM and Cosmos TX differently
+			switch messageType {
+			case EVM, ERC20:
+				signedTx = SignedTx{EvmTx: c.generateSignedEvmTx(keyIndex, messageType)}
+			default:
+				signedTx = SignedTx{TxBytes: c.generateSignedCosmosTxs(keyIndex, messageType, producedCount)}
+			}
+
 			select {
-			case txQueue <- txBytes:
-				atomic.AddInt64(producedCount, 1)
+			case txQueue <- signedTx:
+				producedCount.Add(1)
 			case <-done:
-				// Exit if done signal is received while trying to send to txQueue
 				return
 			}
 		}
 	}
 }
 
-func (c *LoadTestClient) SendTxs(txQueue <-chan []byte, done <-chan struct{}, sentCount *int64, rateLimit int, wg *sync.WaitGroup) {
+func (c *LoadTestClient) generateSignedEvmTx(keyIndex int, msgType string) *ethtypes.Transaction {
+	return c.EvmTxClients[keyIndex].GetTxForMsgType(msgType)
+}
+
+func (c *LoadTestClient) generateSignedCosmosTxs(keyIndex int, msgType string, producedCount *atomic.Int64) []byte {
+	key := c.AccountKeys[keyIndex]
+	msgs, _, _, gas, fee := c.generateMessage(key, msgType)
+	txBuilder := TestConfig.TxConfig.NewTxBuilder()
+	_ = txBuilder.SetMsgs(msgs...)
+	txBuilder.SetGasLimit(gas)
+	txBuilder.SetFeeAmount([]types.Coin{
+		types.NewCoin("usei", types.NewInt(fee)),
+	})
+	// Use random seqno to get around txs that might already be seen in mempool
+	c.SignerClient.SignTx(c.ChainID, &txBuilder, key, uint64(producedCount.Load()))
+	txBytes, _ := TestConfig.TxConfig.TxEncoder()(txBuilder.GetTx())
+	return txBytes
+}
+
+func (c *LoadTestClient) SendTxs(
+	txQueue chan SignedTx,
+	keyIndex int,
+	done <-chan struct{},
+	sentCount *atomic.Int64,
+	semaphore *semaphore.Weighted,
+	wg *sync.WaitGroup,
+) {
+	wg.Add(1)
+	defer wg.Done()
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	rateLimiter := rate.NewLimiter(rate.Limit(rateLimit), rateLimit)
-	maxConcurrent := rateLimit // Set the maximum number of concurrent SendTx calls
-	sem := semaphore.NewWeighted(int64(maxConcurrent))
-
 	for {
 		select {
 		case <-done:
-			fmt.Printf("Stopping consumers\n")
 			return
 		case tx, ok := <-txQueue:
 			if !ok {
 				fmt.Printf("Stopping consumers\n")
 				return
 			}
-
-			if err := sem.Acquire(ctx, 1); err != nil {
+			// Acquire a semaphore
+			if err := semaphore.Acquire(ctx, 1); err != nil {
 				fmt.Printf("Failed to acquire semaphore: %v", err)
 				break
 			}
-			wg.Add(1)
-			go func(tx []byte) {
-				localCtx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-				defer cancel()
-				defer wg.Done()
-				defer sem.Release(1)
-
-				if err := rateLimiter.Wait(localCtx); err != nil {
-					return
+			if tx.TxBytes != nil && len(tx.TxBytes) > 0 {
+				// Send Cosmos Transactions
+				if SendTx(ctx, tx.TxBytes, typestx.BroadcastMode_BROADCAST_MODE_BLOCK, *c) {
+					sentCount.Add(1)
 				}
-				SendTx(ctx, tx, typestx.BroadcastMode_BROADCAST_MODE_BLOCK, false, *c, sentCount)
-			}(tx)
+			} else if tx.EvmTx != nil {
+				// Send EVM Transactions
+				c.EvmTxClients[keyIndex].SendEvmTx(tx.EvmTx, func() {
+					sentCount.Add(1)
+				})
+			}
+			// Release the semaphore
+			semaphore.Release(1)
 		}
 	}
 }
 
+//nolint:staticcheck
 func (c *LoadTestClient) GetTxClient() typestx.ServiceClient {
+	numClients := len(c.TxClients)
+	if numClients <= 0 {
+		panic("There's no Tx client available, make sure your connection are valid")
+	}
 	rand.Seed(time.Now().Unix())
 	return c.TxClients[rand.Int()%len(c.TxClients)]
 }
