@@ -3,16 +3,22 @@ package main
 import (
 	"fmt"
 	"math/rand"
+	"strings"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/ethereum/go-ethereum/common"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+
 	"github.com/sei-protocol/sei-chain/utils"
 )
 
 const (
 	Bank                 string = "bank"
+	EVM                  string = "evm"
+	ERC20                string = "erc20"
 	CollectRewards       string = "collect_rewards"
 	DistributeRewards    string = "distribute_rewards"
 	FailureBankMalformed string = "failure_bank_malformed"
@@ -27,24 +33,22 @@ const (
 	WasmMintNft          string = "wasm_mint_nft"
 	Vortex               string = "vortex"
 	WasmInstantiate      string = "wasm_instantiate"
-	WasmOccIteratorWrite string = "wasm_occ_iterator_write"
-	WasmOccIteratorRange string = "wasm_occ_iterator_range"
-	WasmOccParallelWrite string = "wasm_occ_parallel_write"
 )
 
-type WasmIteratorWriteMsg struct {
-	Values [][]uint64 `json:"values"`
+type EVMAddresses struct {
+	ERC20 common.Address
 }
 
 type Config struct {
 	ChainID            string                `json:"chain_id"`
 	GrpcEndpoints      string                `json:"grpc_endpoints"`
+	EvmRpcEndpoints    string                `json:"evm_rpc_endpoints"`
 	BlockchainEndpoint string                `json:"blockchain_endpoint"`
 	NodeURI            string                `json:"node_uri"`
 	TargetTps          uint64                `json:"target_tps"`
+	MaxAccounts        uint64                `json:"max_accounts"`
 	MsgsPerTx          uint64                `json:"msgs_per_tx"`
-	MessageTypes       []string              `json:"message_types"`
-	RunOracle          bool                  `json:"run_oracle"`
+	MessageType        string                `json:"message_type"`
 	PriceDistr         NumericDistribution   `json:"price_distribution"`
 	QuantityDistr      NumericDistribution   `json:"quantity_distribution"`
 	MsgTypeDistr       MsgTypeDistribution   `json:"message_type_distribution"`
@@ -53,7 +57,26 @@ type Config struct {
 	PerMessageConfigs  MessageConfigs        `json:"message_configs"`
 	MetricsPort        uint64                `json:"metrics_port"`
 	TLS                bool                  `json:"tls"`
-	SeiTesterAddress   string                `json:"sei_tester_address"`
+
+	// These are dynamically set at startup
+	EVMAddresses *EVMAddresses
+}
+
+func (c *Config) EVMRpcEndpoint() string {
+	endpoints := strings.Split(c.EvmRpcEndpoints, ",")
+	return endpoints[0]
+}
+
+func (c *Config) ContainsAnyMessageTypes(types ...string) bool {
+	mTypes := strings.Split(c.MessageType, ",")
+	for _, t := range types {
+		for _, mt := range mTypes {
+			if mt == t {
+				return true
+			}
+		}
+	}
+	return false
 }
 
 type EncodingConfig struct {
@@ -175,4 +198,9 @@ type VortexContract struct {
 type WasmInstantiateType struct {
 	CodeID  uint64 `json:"code_id"`
 	Payload string `json:"payload"`
+}
+
+type SignedTx struct {
+	TxBytes []byte
+	EvmTx   *ethtypes.Transaction
 }
